+++ packages/docusaurus-theme-classic/src/theme/Navbar/index.js
@@ -5,7 +5,7 @@
  * LICENSE file in the root directory of this source tree.
  */
 
-import React, {useCallback, useState, useEffect} from 'react';
+import React, {useCallback, useState} from 'react';
 import Link from '@docusaurus/Link';
 import Head from '@docusaurus/Head';
 import useDocusaurusContext from '@docusaurus/useDocusaurusContext';
@@ -17,6 +17,7 @@ import Toggle from '@theme/Toggle';
 import classnames from 'classnames';
 
 import useTheme from '@theme/hooks/useTheme';
+import useHideableNavbar from '@theme/hooks/useHideableNavbar';
 
 import styles from './styles.module.css';
 
@@ -50,14 +51,8 @@ function Navbar() {
   const [sidebarShown, setSidebarShown] = useState(false);
   const [isSearchBarExpanded, setIsSearchBarExpanded] = useState(false);
   const [theme, setTheme] = useTheme();
-  const [isNavbarVisible, setIsNavbarVisible] = useState(true);
-  const [lastScrollTop, setLastScrollTop] = useState(0);
-  const [navbarHeight, setNavbarHeight] = useState(0);
-  const navbarRef = useCallback(node => {
-    if (node !== null) {
-      setNavbarHeight(node.getBoundingClientRect().height);
-    }
-  }, []);
+
+  const {navbarRef, isNavbarVisible} = useHideableNavbar(hideOnScroll);
 
   const showSidebar = useCallback(() => {
     setSidebarShown(true);
@@ -71,36 +66,6 @@ function Navbar() {
     [setTheme],
   );
 
-  const handleScroll = () => {
-    const scrollTop = window.pageYOffset || document.documentElement.scrollTop;
-    const documentHeight = document.documentElement.scrollHeight - navbarHeight;
-    const windowHeight = window.innerHeight;
-
-    if (scrollTop < navbarHeight) {
-      return;
-    }
-
-    if (lastScrollTop && scrollTop > lastScrollTop) {
-      setIsNavbarVisible(false);
-    } else if (scrollTop + windowHeight < documentHeight) {
-      setIsNavbarVisible(true);
-    }
-
-    setLastScrollTop(scrollTop);
-  };
-
-  useEffect(() => {
-    if (!hideOnScroll) {
-      return undefined;
-    }
-
-    window.addEventListener('scroll', handleScroll);
-
-    return () => {
-      window.removeEventListener('scroll', handleScroll);
-    };
-  }, [lastScrollTop, navbarHeight]);
-
   const logoUrl = useBaseUrl(logo.src);
   return (
     <>
@@ -112,8 +77,8 @@ function Navbar() {
         ref={navbarRef}
         className={classnames('navbar', 'navbar--light', 'navbar--fixed-top', {
           'navbar-sidebar--show': sidebarShown,
-          [styles.navbarHidable]: hideOnScroll,
-          [styles.navbarHided]: !isNavbarVisible,
+          [styles.navbarHideable]: hideOnScroll,
+          [styles.navbarHidden]: !isNavbarVisible,
         })}>
         <div className="navbar__inner">
           <div className="navbar__items">
